+++ Engine/Physic/CollisionDetectionInterface.cpp
@@ -1,5 +1,5 @@
 #include "CollisionDetectionInterface.h"
 
-void Engine::Physic::CollisionDetectionInterface::detection(double deltaTime, std::shared_ptr<std::vector<PhysicObject*>> objects)
+void Engine::Physic::CollisionDetectionInterface::detection(float deltaTime, std::shared_ptr<std::vector<PhysicObject*>> objects)
 {
 }
