+++ Engine/Physic/OctTreeCollisionDetection.cpp
@@ -1,6 +1,6 @@
 #include "OctTreeCollisionDetection.h"
 
-void Engine::Physic::OctTreeCollisionDetection::detection(double deltaTime, std::shared_ptr<std::vector<PhysicObject*>> objects)
+void Engine::Physic::OctTreeCollisionDetection::detection(float deltaTime, std::shared_ptr<std::vector<PhysicObject*>> objects)
 {
 	tree->update();
 	tree->checkCollisions();
